refactor(consumer): update struct definitions for consumer command handling
- Modified Consumer-related structs to improve clarity and consistency
- Updated field names and types for better command parsing and execution
- Adjusted related methods to align with new struct layout

+++ rafka_kafka/src/log/message.rs
@@ -3,7 +3,7 @@ use serde::{Serialize, Deserialize};
 #[derive(Debug, Serialize, Deserialize)]
 
 pub struct Message{
-    pub offset : u64,
+    pub offset : usize,
     pub key : Option<String>,
     pub value : String
 }
